[REVERT]: Remove copying of 370710018002030d00.seq fw
During BT hardware init somehow it finds this fw from system/etc/firmware
and tries to load this and times out. With removing copy of this fw to
system/etc/firmware from btfw.mk, makes BT hw init skips this and
successfully intializes the BT hw.

This need to be reverted back once we resolve this issue.

Change-Id: I6f9abea6aa7f9e6a1662af6a8bcde8b9e99f9647
Signed-off-by: Mantesh sarasetti <manteshx.sarasetti@intel.com>
Reviewed-on: https://android.intel.com/286093
Reviewed-by: sys_pccgand <sys_pccgand@intel.com>
Reviewed-by: Kandru, Suneel <suneel.kandru@intel.com>
Tested-by: Kandru, Suneel <suneel.kandru@intel.com>